api_projects/Medi_Project/rest_auth_project/backend/requirements.txt
django==3.1.2
djangorestframework==3.12.1
django-allauth==0.42.0
django-rest-auth==0.9.5

